% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{The FizzBuzz Function}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{Any  number greater than 0}
}
\value{
The word that was assigned to the number depending on its divisors.
}
\description{
A function that takes a number (n) and prints "Fizz" for a number
divisible by 3, "Buzz" for a number divisible by 5 or "Fizz-Buzz"
for a number divisible by both 3 and 5 for the number of divisors
of the number inputted.
}
\examples{
FizzBuzz(45)

FizzBuzz(20)


}
