% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather Information from Flagstaff Arizona from 2015 to 2019.}
\format{
a data frame with 1719 observations of 4 variables
\describe{
  \item{Date}{The date of observation in Year-Month-Day date format.}
  \item{PRCP}{The amount of Precipitation in Inches.}
  \item{SNOW}{The amount of Snowfall in Inches}
  \item{TMAX}{The maximum temperature in Fahrenheit.}
}
}
\source{
\url{https://www.ncdc.noaa.gov”}
}
\usage{
Flagstaff_Weather
}
\description{
Weather Information from Flagstaff Arizona from 2015 to 2019.
}
\keyword{datasets}
